Update main.py
cleaned up code, extensive comments to help understand what exactly is going on with each line. an AI helped me get a prototype going a few months ago, now its time to take this shit to the next level.

What this really represents is the first of many functions to be built. It is a simple squared symmetry transformation of the original inputted image.